Move Platform class into preprocessor.py
There was a circular dependency between platform.py and preprocessor.py:
- Platform stores Macro and MacroFunction objects; and
- MacroExpander requires a Platform.

Breaking this dependency enables Platform arguments in preprocessor.py to be
decorated with the correct type-hints. This code structure is also a better
reflection of how Platform is used by the preprocessor.
