+++ tests/operators/test_operators.py
@@ -5,7 +5,8 @@ import logging
 import unittest
 from pathlib import Path
 
-from codebasin import CodeBase, finder, platform, preprocessor
+from codebasin import CodeBase, finder, preprocessor
+from codebasin.preprocessor import Platform
 
 
 class TestOperators(unittest.TestCase):
@@ -52,7 +53,7 @@ class TestOperators(unittest.TestCase):
     def test_paths(self):
         input_str = r"FUNCTION(looks/2like/a/path/with_/bad%%identifiers)"
         tokens = preprocessor.Lexer(input_str).tokenize()
-        p = platform.Platform("Test", self.rootdir)
+        p = Platform("Test", self.rootdir)
         macro = preprocessor.macro_from_definition_string("FUNCTION(x)=#x")
         p._definitions = {macro.name: macro}
         _ = preprocessor.MacroExpander(p).expand(tokens)
